array is no longer a uint8_t

+++ Core/Inc/neom9n.h
@@ -85,7 +85,7 @@ typedef struct {
 // Function prototypes
 void gps_init(NEOM9N_t *packet, SPI_HandleTypeDef *hspi);
 bool gps_has_fix(const NEOM9N_t *packet);
-void pack_gps_data(const NEOM9N_t *packet, uint8_t *buffer);
+void pack_gps_data(const NEOM9N_t *packet, float *buffer);
 uint64_t get_datetime_data(const NEOM9N_t *packet);
 NEOM9N_status_t read_nmea_gga(NEOM9N_t *packet, uint32_t max_wait);
 NEOM9N_status_t read_nmea_rmc(NEOM9N_t *packet, uint32_t max_wait);
